Merge branch 'master' into openmp

# Conflicts:
#	R/RcppExports.R
#	src/Rtsne.cpp
#	src/tsne.cpp

+++ src/tsne.cpp
@@ -112,21 +112,21 @@ void TSNE<NDims>::run(double* X, int N, int D, double* Y, int no_dims,
     if(exact) {
         
         // Compute similarities
-        P = (double*) malloc(N * N * sizeof(double));
+        P = (double*) malloc((long)N * N * sizeof(double));
         if(P == NULL) { Rcpp::stop("Memory allocation failed!\n"); }
         computeGaussianPerplexity(X, N, D, P, perplexity, distance_precomputed);
     
         // Symmetrize input similarities
         if (verbose) Rprintf("Symmetrizing...\n");
-        for(int n = 0; n < N; n++) {
-            for(int m = n + 1; m < N; m++) {
+        for(unsigned long n = 0; n < N; n++) {
+            for(unsigned long m = n + 1; m < N; m++) {
                 P[n * N + m] += P[m * N + n];
                 P[m * N + n]  = P[n * N + m];
             }
         }
         double sum_P = .0;
-        for(int i = 0; i < N * N; i++) sum_P += P[i];
-        for(int i = 0; i < N * N; i++) P[i] /= sum_P;
+        for(unsigned long i = 0; i < (long)N * N; i++) sum_P += P[i];
+        for(unsigned long i = 0; i < (long)N * N; i++) P[i] /= sum_P;
     }
     
     // Compute input similarities for approximate t-SNE
@@ -150,8 +150,8 @@ void TSNE<NDims>::run(double* X, int N, int D, double* Y, int no_dims,
     end = clock();
     
     // Lie about the P-values
-    if(exact) { for(int i = 0; i < N * N; i++)        P[i] *= exaggeration_factor; }
-    else {      for(int i = 0; i < row_P[N]; i++) val_P[i] *= exaggeration_factor; }
+    if(exact) { for(unsigned long i = 0; i < (long)N * N; i++)        P[i] *= exaggeration_factor; }
+    else {      for(unsigned long i = 0; i < row_P[N]; i++) val_P[i] *= exaggeration_factor; }
 
 	// Initialize solution (randomly), if not already done
 	if (!init) { for(int i = 0; i < N * no_dims; i++) Y[i] = randn() * .0001; }
@@ -169,7 +169,7 @@ void TSNE<NDims>::run(double* X, int N, int D, double* Y, int no_dims,
         
         // Stop lying about the P-values after a while, and switch momentum
         if(iter == stop_lying_iter) {
-          if(exact) { for(int i = 0; i < N * N; i++)        P[i] /= exaggeration_factor; }
+          if(exact) { for(unsigned long i = 0; i < (long)N * N; i++)        P[i] /= exaggeration_factor; }
           else      { for(int i = 0; i < row_P[N]; i++) val_P[i] /= exaggeration_factor; }
         }
         if(iter == mom_switch_iter) momentum = final_momentum;
@@ -264,16 +264,16 @@ void TSNE<NDims>::computeExactGradient(double* P, double* Y, int N, int D, doubl
 	for(int i = 0; i < N * D; i++) dC[i] = 0.0;
     
     // Compute the squared Euclidean distance matrix
-    double* DD = (double*) malloc(N * N * sizeof(double));
+    double* DD = (double*) malloc((long)N * N * sizeof(double));
     if(DD == NULL) { Rcpp::stop("Memory allocation failed!\n"); }
     computeSquaredEuclideanDistance(Y, N, D, DD);
     
     // Compute Q-matrix and normalization sum
-    double* Q    = (double*) malloc(N * N * sizeof(double));
+    double* Q    = (double*) malloc((long)N * N * sizeof(double));
     if(Q == NULL) { Rcpp::stop("Memory allocation failed!\n"); }
     double sum_Q = .0;
-    for(int n = 0; n < N; n++) {
-    	for(int m = 0; m < N; m++) {
+    for(unsigned long n = 0; n < N; n++) {
+    	for(unsigned long m = 0; m < N; m++) {
             if(n != m) {
                 Q[n * N + m] = 1 / (1 + DD[n * N + m]);
                 sum_Q += Q[n * N + m];
@@ -282,8 +282,8 @@ void TSNE<NDims>::computeExactGradient(double* P, double* Y, int N, int D, doubl
     }
     
 	// Perform the computation of the gradient
-	for(int n = 0; n < N; n++) {
-    	for(int m = 0; m < N; m++) {
+	for(unsigned long n = 0; n < N; n++) {
+    	for(unsigned long m = 0; m < N; m++) {
             if(n != m) {
                 double mult = (P[n * N + m] - (Q[n * N + m] / sum_Q)) * Q[n * N + m];
                 for(int d = 0; d < D; d++) {
@@ -304,15 +304,15 @@ template <int NDims>
 double TSNE<NDims>::evaluateError(double* P, double* Y, int N, int D) {
     
     // Compute the squared Euclidean distance matrix
-    double* DD = (double*) malloc(N * N * sizeof(double));
-    double* Q = (double*) malloc(N * N * sizeof(double));
+    double* DD = (double*) malloc((long)N * N * sizeof(double));
+    double* Q = (double*) malloc((long)N * N * sizeof(double));
     if(DD == NULL || Q == NULL) { Rcpp::stop("Memory allocation failed!\n"); }
     computeSquaredEuclideanDistance(Y, N, D, DD);
     
     // Compute Q-matrix and normalization sum
     double sum_Q = DBL_MIN;
-    for(int n = 0; n < N; n++) {
-    	for(int m = 0; m < N; m++) {
+    for(unsigned long n = 0; n < N; n++) {
+    	for(unsigned long m = 0; m < N; m++) {
             if(n != m) {
                 Q[n * N + m] = 1 / (1 + DD[n * N + m]);
                 sum_Q += Q[n * N + m];
@@ -320,15 +320,15 @@ double TSNE<NDims>::evaluateError(double* P, double* Y, int N, int D) {
             else Q[n * N + m] = DBL_MIN;
         }
     }
-    for(int i = 0; i < N * N; i++) Q[i] /= sum_Q;
+    for(unsigned long i = 0; i < (long)N * N; i++) Q[i] /= sum_Q;
     
     // Sum t-SNE error
     double C = .0;
-	for(int n = 0; n < N; n++) {
-		for(int m = 0; m < N; m++) {
-			C += P[n * N + m] * log((P[n * N + m] + 1e-9) / (Q[n * N + m] + 1e-9));
-		}
-	}
+  	for(unsigned long n = 0; n < N; n++) {
+  		for(unsigned long m = 0; m < N; m++) {
+  			C += P[n * N + m] * log((P[n * N + m] + 1e-9) / (Q[n * N + m] + 1e-9));
+  		}
+  	}
     
     // Clean up memory
     free(DD);
@@ -374,15 +374,15 @@ template <int NDims>
 void TSNE<NDims>::getCost(double* P, double* Y, int N, int D, double* costs) {
   
   // Compute the squared Euclidean distance matrix
-  double* DD = (double*) malloc(N * N * sizeof(double));
-  double* Q = (double*) malloc(N * N * sizeof(double));
+  double* DD = (double*) malloc((long)N * N * sizeof(double));
+  double* Q = (double*) malloc((long)N * N * sizeof(double));
   if(DD == NULL || Q == NULL) { Rcpp::stop("Memory allocation failed!\n"); }
   computeSquaredEuclideanDistance(Y, N, D, DD);
   
   // Compute Q-matrix and normalization sum
   double sum_Q = DBL_MIN;
-  for(int n = 0; n < N; n++) {
-    for(int m = 0; m < N; m++) {
+  for(unsigned long n = 0; n < N; n++) {
+    for(unsigned long m = 0; m < N; m++) {
       if(n != m) {
         Q[n * N + m] = 1 / (1 + DD[n * N + m]);
         sum_Q += Q[n * N + m];
@@ -390,12 +390,12 @@ void TSNE<NDims>::getCost(double* P, double* Y, int N, int D, double* costs) {
       else Q[n * N + m] = DBL_MIN;
     }
   }
-  for(int i = 0; i < N * N; i++) Q[i] /= sum_Q;
+  for(unsigned long i = 0; i < (long)N * N; i++) Q[i] /= sum_Q;
   
   // Sum t-SNE error
-  for(int n = 0; n < N; n++) {
+  for(unsigned long n = 0; n < N; n++) {
     costs[n] = 0.0;
-    for(int m = 0; m < N; m++) {
+    for(unsigned long m = 0; m < N; m++) {
       costs[n] += P[n * N + m] * log((P[n * N + m] + 1e-9) / (Q[n * N + m] + 1e-9));
     }
   }
@@ -444,7 +444,7 @@ template <int NDims>
 void TSNE<NDims>::computeGaussianPerplexity(double* X, int N, int D, double* P, double perplexity, bool distance_precomputed) {
 	
 	// Compute the squared Euclidean distance matrix
-	double* DD = (double*) malloc(N * N * sizeof(double));
+	double* DD = (double*) malloc((long)N * N * sizeof(double));
   if(DD == NULL) { Rcpp::stop("Memory allocation failed!\n"); }
 	
 	if (distance_precomputed) {
@@ -459,7 +459,7 @@ void TSNE<NDims>::computeGaussianPerplexity(double* X, int N, int D, double* P, 
 // 	}
 
 	// Compute the Gaussian kernel row by row
-	for(int n = 0; n < N; n++) {
+	for(unsigned long n = 0; n < N; n++) {
         
 		// Initialize some variables
 		bool found = false;
@@ -474,14 +474,14 @@ void TSNE<NDims>::computeGaussianPerplexity(double* X, int N, int D, double* P, 
 		while(!found && iter < 200) {
 			
 			// Compute Gaussian kernel row
-			for(int m = 0; m < N; m++) P[n * N + m] = exp(-beta * DD[n * N + m]);
+			for(unsigned long m = 0; m < N; m++) P[n * N + m] = exp(-beta * DD[n * N + m]);
 			P[n * N + n] = DBL_MIN;
 			
 			// Compute entropy of current row
 			sum_P = DBL_MIN;
-			for(int m = 0; m < N; m++) sum_P += P[n * N + m];
+			for(unsigned long m = 0; m < N; m++) sum_P += P[n * N + m];
 			double H = 0.0;
-			for(int m = 0; m < N; m++) H += beta * (DD[n * N + m] * P[n * N + m]);
+			for(unsigned long m = 0; m < N; m++) H += beta * (DD[n * N + m] * P[n * N + m]);
 			H = (H / sum_P) + log(sum_P);
 			
 			// Evaluate whether the entropy is within the tolerance level
@@ -511,7 +511,7 @@ void TSNE<NDims>::computeGaussianPerplexity(double* X, int N, int D, double* P, 
 		}
 		
 		// Row normalize P
-		for(int m = 0; m < N; m++) P[n * N + m] /= sum_P;
+		for(unsigned long m = 0; m < N; m++) P[n * N + m] /= sum_P;
 	}
 	
 	// Clean up memory
@@ -723,8 +723,8 @@ void TSNE<NDims>::computeSquaredEuclideanDistance(double* X, int N, int D, doubl
             dataSums[n] += (X[n * D + d] * X[n * D + d]);
         }
     }
-    for(int n = 0; n < N; n++) {
-        for(int m = 0; m < N; m++) {
+    for(unsigned long n = 0; n < N; n++) {
+        for(unsigned long m = 0; m < N; m++) {
             DD[n * N + m] = dataSums[n] + dataSums[m];
         }
     }
